mutt: Default configuration for my gmail
For now, I only use this to finish The Eudyptula challenge. Afterwards, if I
need to use something like mutt, I will do it through the uwaterloo account.
This is because I haven't found a configuration for Gmail that works without
turning on:

``Allowing less secure apps to access your account''
https://support.google.com/accounts/answer/6010255?hl=en

I will manually switch it on and off for when I have to submit the eudyptula
code, to get some confirmation that my submission went through.